backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
Pillow==10.4.0
SQLAlchemy==2.0.34

# Machine Learning dependencies
transformers==4.36.0
torch==2.1.0
torchvision==0.16.0

